+++ bindings/go/oci/internal/attestation/constants.go
